// src/lib/pwa/types.ts
export interface BeforeInstallPromptEvent extends Event {
  readonly platforms: string[]
  readonly userChoice: Promise<{
    outcome: "accepted" | "dismissed"
    platform: string
  }>
  prompt: () => Promise<void>
}

export interface NavigatorWithStandalone extends Navigator {
  standalone?: boolean
}